Consistent file naming for specs (and *TernaryOperator cops)
