+++ lib/rubocop.rb
@@ -84,6 +84,7 @@ require 'rubocop/cop/style/begin_block'
 require 'rubocop/cop/style/block_comments'
 require 'rubocop/cop/style/block_nesting'
 require 'rubocop/cop/style/blocks'
+require 'rubocop/cop/style/braces_around_hash_parameters'
 require 'rubocop/cop/style/case_equality'
 require 'rubocop/cop/style/case_indentation'
 require 'rubocop/cop/style/character_literal'
@@ -101,8 +102,8 @@ require 'rubocop/cop/style/documentation'
 require 'rubocop/cop/style/dot_position'
 require 'rubocop/cop/style/empty_line_between_defs'
 require 'rubocop/cop/style/empty_lines'
-require 'rubocop/cop/style/empty_lines_around_body'
 require 'rubocop/cop/style/empty_lines_around_access_modifier'
+require 'rubocop/cop/style/empty_lines_around_body'
 require 'rubocop/cop/style/empty_literal'
 require 'rubocop/cop/style/encoding'
 require 'rubocop/cop/style/end_block'
@@ -125,14 +126,16 @@ require 'rubocop/cop/style/lambda'
 require 'rubocop/cop/style/lambda_call'
 require 'rubocop/cop/style/leading_comment_space'
 require 'rubocop/cop/style/line_length'
-require 'rubocop/cop/style/method_def_parentheses'
 require 'rubocop/cop/style/method_call_parentheses'
 require 'rubocop/cop/style/method_called_on_do_end_block'
+require 'rubocop/cop/style/method_def_parentheses'
 require 'rubocop/cop/style/method_length'
 require 'rubocop/cop/style/method_name'
 require 'rubocop/cop/style/module_function'
 require 'rubocop/cop/style/multiline_block_chain'
 require 'rubocop/cop/style/multiline_if_then'
+require 'rubocop/cop/style/multiline_ternary_operator'
+require 'rubocop/cop/style/nested_ternary_operator'
 require 'rubocop/cop/style/nil_comparison'
 require 'rubocop/cop/style/not'
 require 'rubocop/cop/style/numeric_literals'
@@ -164,14 +167,13 @@ require 'rubocop/cop/style/space_around_block_braces'
 require 'rubocop/cop/style/space_around_equals_in_parameter_default'
 require 'rubocop/cop/style/space_around_operators'
 require 'rubocop/cop/style/space_before_modifier_keyword'
-require 'rubocop/cop/style/space_inside_hash_literal_braces'
 require 'rubocop/cop/style/space_inside_brackets'
+require 'rubocop/cop/style/space_inside_hash_literal_braces'
 require 'rubocop/cop/style/space_inside_parens'
 require 'rubocop/cop/style/special_global_vars'
 require 'rubocop/cop/style/string_literals'
 require 'rubocop/cop/style/symbol_array'
 require 'rubocop/cop/style/tab'
-require 'rubocop/cop/style/ternary_operator'
 require 'rubocop/cop/style/trailing_blank_lines'
 require 'rubocop/cop/style/trailing_whitespace'
 require 'rubocop/cop/style/trivial_accessors'
@@ -182,7 +184,6 @@ require 'rubocop/cop/style/when_then'
 require 'rubocop/cop/style/while_until_do'
 require 'rubocop/cop/style/while_until_modifier'
 require 'rubocop/cop/style/word_array'
-require 'rubocop/cop/style/braces_around_hash_parameters'
 
 require 'rubocop/cop/rails/default_scope'
 require 'rubocop/cop/rails/has_and_belongs_to_many'
