+++ spec/rubocop/cop/style/space_around_equals_in_parameter_default_spec.rb
